click:  align Ipv4L3ClickProtocol with ae25cc3a0f89
this commit also fixes a bug uncovered by the more strict parsing
of Ipv4Address strings.  In Ipv4ClickRouting::RouteOutput(), it is
possible for the string from the click routing table to contain
just a single entry, but the code was assuming a space-delimited
pair.  If only a single entry is returned, the destination is now
set to "0.0.0.0".