wifi:  Remove Simulator::Run () calls in the PARF unit tests
The problem that may arise is that the unit test performs a number of
operations on Wifi objects before Simulator::Run() is entered.  There
are no events generated by these tests, so Simulator::Run () will not do
anything for the test.  What it may do, however, is call some
Object::DoInitialize() code that may assert if it finds the Wifi object
in an unexpected state.